% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data__models.R
\name{aaipi}
\alias{aaipi}
\title{The Age Adjusted International Prognostic Index (aaIPI)}
\format{
A named vector of length 3. See Examples for details
}
\source{
\url{https://doi.org/10.1056/NEJM199309303291402}
}
\description{
The Age Adjusted International Prognostic Index (aaIPI) as
described in \emph{A predictive model for aggressive non-Hodgkin's lymphoma. The
International Non-Hodgkin's Lymphoma Prognostic Factors Project. N Engl J
Med 1993;329(14):987-94}.

The aaIPI works by assigning one point for each the following factors:
\itemize{
\item Stage III or IV disease
\item Elevated serum LDH
\item ECOG/Zubrod performance status of 2, 3, or 4
}

The sum of points can be used to separate samples into following risk groups:
\itemize{
\item Low risk (0 points)
\item Low-intermediate risk (1 points)
\item High-intermediate risk (2 points)
\item High risk (3 points)
}
}
\examples{
\dontrun{
as.data.frame(aaipi)
#                   ipi
# stage_3_4           1  # Stage III or IV disease
# lhd_elevated        1  # Elevated serum LDH
# ecog_2_3_4          1  # ECOG/Zubrod performance status of 2, 3, or 4
}
}
