% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/071_act_set_auth_state_if_returning_from_github_login.R
\name{act_set_auth_state_if_returning_from_github_login}
\alias{act_set_auth_state_if_returning_from_github_login}
\title{Set auth state based on Github OAuth Token}
\usage{
act_set_auth_state_if_returning_from_github_login(ses)
}
\arguments{
\item{ses}{List of session specific objects as returned by \code{init_data} once
for every user session}
}
\description{
Set auth state based on Github OAuth Token
}
\details{
Should be called once at the beginning of the server function
}
