% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/010_cli.R
\name{cli}
\alias{cli}
\title{Commandline Interface}
\usage{
cli(argv = commandArgs(trailingOnly = TRUE))
}
\arguments{
\item{argv}{character vector of commandline arguments}
}
\description{
Entry point when called directly from a system shell such as
bash. I.e. calling imlui:::cli("--help") from an interactive R session is
equivalent to calling \verb{R --no-echo --vanilla -q -e "imlui:::cli()" --help}
from bash.
}
\details{
To make shell usage more convenient, it is strongly recommended to
create an alias/wrapper for your shell of choice. To create an alias in
bash, add the line 'alias imlui=R --no-echo --vanilla -q -e "imlui:::cli()"'
to ~/.bashrc. To create an alias for CMD.exe, create a file "imlui.bat" in a
folder added to the PATH environment variable containing the line 'R
--no-echo --vanilla -q -e "imlui:::cli()" \%*'.

Now \code{imlui} can be used directly from commandline as in \code{imlui --help}.
}
\examples{
sum(1:10); \dontrun{sum('a', 'b')}
cli(c("runserver"))
cli(c("--help"))
}
