% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/020_db.R
\name{DB}
\alias{DB}
\title{R6 class representing a database}
\description{
R6 class representing a database

R6 class representing a database
}
\examples{

## ------------------------------------------------
## Method `DB$execute`
## ------------------------------------------------

\dontrun{
  execute("INSERT INTO cars (speed, dist) VALUES (1, 1), (2, 2), (3, 3)")
  execute(
    query = "UPDATE Appstate SET resource_value = ? WHERE
             user_id = ? AND resource_id = ?",
    params = list(url, user_id, "url"))
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{type}}{todo}

\item{\code{filepath}}{todo}

\item{\code{hostname}}{todo}

\item{\code{port}}{todo}

\item{\code{database}}{todo}

\item{\code{username}}{todo}

\item{\code{password}}{todo}

\item{\code{conn}}{todo}

\item{\code{default_filepath}}{todo}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-DB-new}{\code{DB$new()}}
\item \href{#method-DB-connect}{\code{DB$connect()}}
\item \href{#method-DB-disconnect}{\code{DB$disconnect()}}
\item \href{#method-DB-is_connected}{\code{DB$is_connected()}}
\item \href{#method-DB-finalize}{\code{DB$finalize()}}
\item \href{#method-DB-get_column}{\code{DB$get_column()}}
\item \href{#method-DB-get_colnames}{\code{DB$get_colnames()}}
\item \href{#method-DB-get_table}{\code{DB$get_table()}}
\item \href{#method-DB-get_tables}{\code{DB$get_tables()}}
\item \href{#method-DB-execute}{\code{DB$execute()}}
\item \href{#method-DB-get_valid_cookies}{\code{DB$get_valid_cookies()}}
\item \href{#method-DB-insert_cookie}{\code{DB$insert_cookie()}}
\item \href{#method-DB-appendTable}{\code{DB$appendTable()}}
\item \href{#method-DB-canConnect}{\code{DB$canConnect()}}
\item \href{#method-DB-createTable}{\code{DB$createTable()}}
\item \href{#method-DB-existsTable}{\code{DB$existsTable()}}
\item \href{#method-DB-fetch}{\code{DB$fetch()}}
\item \href{#method-DB-getInfo}{\code{DB$getInfo()}}
\item \href{#method-DB-getRowCount}{\code{DB$getRowCount()}}
\item \href{#method-DB-getRowsAffected}{\code{DB$getRowsAffected()}}
\item \href{#method-DB-getStatement}{\code{DB$getStatement()}}
\item \href{#method-DB-hasCompleted}{\code{DB$hasCompleted()}}
\item \href{#method-DB-isReadOnly}{\code{DB$isReadOnly()}}
\item \href{#method-DB-listObjects}{\code{DB$listObjects()}}
\item \href{#method-DB-listResults}{\code{DB$listResults()}}
\item \href{#method-DB-listTables}{\code{DB$listTables()}}
\item \href{#method-DB-quoteIdentifier}{\code{DB$quoteIdentifier()}}
\item \href{#method-DB-quoteLiteral}{\code{DB$quoteLiteral()}}
\item \href{#method-DB-quoteString}{\code{DB$quoteString()}}
\item \href{#method-DB-readTable}{\code{DB$readTable()}}
\item \href{#method-DB-removeTable}{\code{DB$removeTable()}}
\item \href{#method-DB-rollback}{\code{DB$rollback()}}
\item \href{#method-DB-sendQuery}{\code{DB$sendQuery()}}
\item \href{#method-DB-sendStatement}{\code{DB$sendStatement()}}
\item \href{#method-DB-setDataMappings}{\code{DB$setDataMappings()}}
\item \href{#method-DB-writeTable}{\code{DB$writeTable()}}
\item \href{#method-DB-clone}{\code{DB$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-new"></a>}}
\if{latex}{\out{\hypertarget{method-DB-new}{}}}
\subsection{Method \code{new()}}{
Create and connect a new DB object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$new(config = pkg$imlui_config$dbms)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{Either list(type="sqlite", filepath="...") or
list(type="postgres", hostname="...", port="...", database="...",
username="...", password="...")}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{DB} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-connect"></a>}}
\if{latex}{\out{\hypertarget{method-DB-connect}{}}}
\subsection{Method \code{connect()}}{
Connect to database
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$connect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-disconnect"></a>}}
\if{latex}{\out{\hypertarget{method-DB-disconnect}{}}}
\subsection{Method \code{disconnect()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$disconnect()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-is_connected"></a>}}
\if{latex}{\out{\hypertarget{method-DB-is_connected}{}}}
\subsection{Method \code{is_connected()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$is_connected()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-finalize"></a>}}
\if{latex}{\out{\hypertarget{method-DB-finalize}{}}}
\subsection{Method \code{finalize()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$finalize()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-get_column"></a>}}
\if{latex}{\out{\hypertarget{method-DB-get_column}{}}}
\subsection{Method \code{get_column()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$get_column(col, tbl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{col}}{column name}

\item{\code{tbl}}{table name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-get_colnames"></a>}}
\if{latex}{\out{\hypertarget{method-DB-get_colnames}{}}}
\subsection{Method \code{get_colnames()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$get_colnames(tbl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl}}{table name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-get_table"></a>}}
\if{latex}{\out{\hypertarget{method-DB-get_table}{}}}
\subsection{Method \code{get_table()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$get_table(tbl)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tbl}}{table name}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-get_tables"></a>}}
\if{latex}{\out{\hypertarget{method-DB-get_tables}{}}}
\subsection{Method \code{get_tables()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$get_tables()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-execute"></a>}}
\if{latex}{\out{\hypertarget{method-DB-execute}{}}}
\subsection{Method \code{execute()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$execute(query, params = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{query}}{SQL query as charactern string (can include ?)}

\item{\code{params}}{parameters for query used as substitute for ? in \code{query}}
}
\if{html}{\out{</div>}}
}
\subsection{Examples}{
\if{html}{\out{<div class="r example copy">}}
\preformatted{\dontrun{
  execute("INSERT INTO cars (speed, dist) VALUES (1, 1), (2, 2), (3, 3)")
  execute(
    query = "UPDATE Appstate SET resource_value = ? WHERE
             user_id = ? AND resource_id = ?",
    params = list(url, user_id, "url"))
}
}
\if{html}{\out{</div>}}

}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-get_valid_cookies"></a>}}
\if{latex}{\out{\hypertarget{method-DB-get_valid_cookies}{}}}
\subsection{Method \code{get_valid_cookies()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$get_valid_cookies(expiry = 7)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{expiry}}{Expiry date in days}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-insert_cookie"></a>}}
\if{latex}{\out{\hypertarget{method-DB-insert_cookie}{}}}
\subsection{Method \code{insert_cookie()}}{
This function must accept two parameters: user_id and
session_id. It will be called whenever the user successfully logs in with
a password. This function saves to your database. TODO: Based on
shinyauthr example (cite!).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$insert_cookie(user_id, session_id)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{user_id}}{user_id}

\item{\code{session_id}}{sessiond_id}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-appendTable"></a>}}
\if{latex}{\out{\hypertarget{method-DB-appendTable}{}}}
\subsection{Method \code{appendTable()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$appendTable(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-canConnect"></a>}}
\if{latex}{\out{\hypertarget{method-DB-canConnect}{}}}
\subsection{Method \code{canConnect()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$canConnect(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-createTable"></a>}}
\if{latex}{\out{\hypertarget{method-DB-createTable}{}}}
\subsection{Method \code{createTable()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$createTable(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-existsTable"></a>}}
\if{latex}{\out{\hypertarget{method-DB-existsTable}{}}}
\subsection{Method \code{existsTable()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$existsTable(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-fetch"></a>}}
\if{latex}{\out{\hypertarget{method-DB-fetch}{}}}
\subsection{Method \code{fetch()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$fetch(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-getInfo"></a>}}
\if{latex}{\out{\hypertarget{method-DB-getInfo}{}}}
\subsection{Method \code{getInfo()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$getInfo(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-getRowCount"></a>}}
\if{latex}{\out{\hypertarget{method-DB-getRowCount}{}}}
\subsection{Method \code{getRowCount()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$getRowCount(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-getRowsAffected"></a>}}
\if{latex}{\out{\hypertarget{method-DB-getRowsAffected}{}}}
\subsection{Method \code{getRowsAffected()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$getRowsAffected(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-getStatement"></a>}}
\if{latex}{\out{\hypertarget{method-DB-getStatement}{}}}
\subsection{Method \code{getStatement()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$getStatement(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-hasCompleted"></a>}}
\if{latex}{\out{\hypertarget{method-DB-hasCompleted}{}}}
\subsection{Method \code{hasCompleted()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$hasCompleted(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-isReadOnly"></a>}}
\if{latex}{\out{\hypertarget{method-DB-isReadOnly}{}}}
\subsection{Method \code{isReadOnly()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$isReadOnly(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-listObjects"></a>}}
\if{latex}{\out{\hypertarget{method-DB-listObjects}{}}}
\subsection{Method \code{listObjects()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$listObjects(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-listResults"></a>}}
\if{latex}{\out{\hypertarget{method-DB-listResults}{}}}
\subsection{Method \code{listResults()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$listResults(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-listTables"></a>}}
\if{latex}{\out{\hypertarget{method-DB-listTables}{}}}
\subsection{Method \code{listTables()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$listTables(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-quoteIdentifier"></a>}}
\if{latex}{\out{\hypertarget{method-DB-quoteIdentifier}{}}}
\subsection{Method \code{quoteIdentifier()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$quoteIdentifier(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-quoteLiteral"></a>}}
\if{latex}{\out{\hypertarget{method-DB-quoteLiteral}{}}}
\subsection{Method \code{quoteLiteral()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$quoteLiteral(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-quoteString"></a>}}
\if{latex}{\out{\hypertarget{method-DB-quoteString}{}}}
\subsection{Method \code{quoteString()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$quoteString(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-readTable"></a>}}
\if{latex}{\out{\hypertarget{method-DB-readTable}{}}}
\subsection{Method \code{readTable()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$readTable(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-removeTable"></a>}}
\if{latex}{\out{\hypertarget{method-DB-removeTable}{}}}
\subsection{Method \code{removeTable()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$removeTable(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-rollback"></a>}}
\if{latex}{\out{\hypertarget{method-DB-rollback}{}}}
\subsection{Method \code{rollback()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$rollback(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-sendQuery"></a>}}
\if{latex}{\out{\hypertarget{method-DB-sendQuery}{}}}
\subsection{Method \code{sendQuery()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$sendQuery(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-sendStatement"></a>}}
\if{latex}{\out{\hypertarget{method-DB-sendStatement}{}}}
\subsection{Method \code{sendStatement()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$sendStatement(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-setDataMappings"></a>}}
\if{latex}{\out{\hypertarget{method-DB-setDataMappings}{}}}
\subsection{Method \code{setDataMappings()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$setDataMappings(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-writeTable"></a>}}
\if{latex}{\out{\hypertarget{method-DB-writeTable}{}}}
\subsection{Method \code{writeTable()}}{
todo
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$writeTable(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{todo}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-DB-clone"></a>}}
\if{latex}{\out{\hypertarget{method-DB-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{DB$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
