% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db__connect}
\alias{db__connect}
\title{Connect to IMLUI_DB.}
\usage{
db__connect(force_recreate = FALSE)
}
\arguments{
\item{force_recreate}{Logical, if TRUE, the database will be recreated from scratch. This is useful for development.}
}
\description{
Connect to database specifed in IMLUI_CONFIG. Resulting connection object is stored in process environment \code{penv}.
}
\details{
Each call to \code{db__connect} should be paired with a call to \code{db__disconnect()}. See Examples for details.
}
\examples{
\dontrun{
db__connect()
on.exit(db__disconnect(), add = TRUE)
}
}
