% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db__get__default_model_selection}
\alias{db__get__default_model_selection}
\title{Get default model selection for current user from database}
\usage{
db__get__default_model_selection(ses, choices = NULL, n = NULL)
}
\arguments{
\item{ses}{Session envrionment}

\item{choices}{Allowed choices}

\item{n}{Maximum number of models to return}
}
\description{
Returns the default model selection for the current user as
character vector. If the setting "models" is not specified in the database
or contains no models available in \code{choices}, NULL is returned. If the
number of models specified (after filtering by \code{choices}) is still greater
than \code{n}, only the first \code{n} models are returned.
}
