% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db__get_colnames}
\alias{db__get_colnames}
\title{Get all column names from one Table of IMLUI_DATABASE}
\usage{
db__get_colnames(tbl)
}
\arguments{
\item{tbl}{String, name of table}
}
\description{
Get column names from table \code{tbl} as vector
}
\examples{
\dontrun{
db__connect()
db__get_colnames("users")
}
}
