% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db__get_column}
\alias{db__get_column}
\title{Get one Column from one Table of IMLUI_DATABASE}
\usage{
db__get_column(col, tbl)
}
\arguments{
\item{col}{String, name of column}

\item{tbl}{String, name of table}
}
\description{
Get column \code{col} from table \code{tbl} as vector.
}
\examples{
\dontrun{
db__connect()
db__get_column("id", tbl = "users")
}
}
