% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db__get_valid_cookies}
\alias{db__get_valid_cookies}
\title{Fetch valid session IDs from IMLUI_DATABASE}
\usage{
db__get_valid_cookies(expiry = 7)
}
\arguments{
\item{expiry}{Expiry date in days}
}
\description{
Get valid, i.e. not yet expired, cookies from Database.
}
