% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{db__insert_cookie}
\alias{db__insert_cookie}
\title{Store user ID and session ID in IMLUI_DATABASE}
\usage{
db__insert_cookie(user_id, session_id)
}
\arguments{
\item{user_id}{user_id}

\item{session_id}{sessiond_id}
}
\description{
This function should be called whenever a user logs in successfully. It stores the provided user ID \code{user_id} together with the current session ID \code{session_id} in table \code{mapping_users_sessions} of IMLUI_DATABASE. It should always be paired with call to \code{set_login_cookie()}. See Examples for an example.
}
\examples{
\dontrun{
db__connect()
db__insert_cookie(user_id = "max", session_id = "asdf1234")
set_login_cookie(ses, user_id = "max", session_id = "asdf1234")
}
}
