% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data__datasets.R
\name{dummy_nordmo_2020_training_dataset}
\alias{dummy_nordmo_2020_training_dataset}
\title{Dummy Dataset}
\format{
See section Details.
}
\source{
\url{https://doi.org/10.1080/10428194.2020.1861268}
}
\description{
This is \strong{not} the training dataset used in Nordmo, Carmen,
Gunther Glehr, Michael Altenbuchinger, Rainer Spang, Marita Ziepert, Heike
Horn, Annette M. Staiger, et al. "Identification of a MiRNA Based Model to
Detect Prognostic Subgroups in Patients with Aggressive B-Cell Lymphoma".
Leukemia & Lymphoma 62, no. 5 (16 April 2021): 1107-15.
URL: \url{https://doi.org/10.1080/10428194.2020.1861268}. Instead, this dataset
was \strong{simulated completely from scratch} to allow for a test of the
imlui application without having access to the real dataset.
}
\examples{
\dontrun{
> head(dummy_nordmo_2020_training_dataset)
  hsa-miR-106b-5p hsa-miR-130a-3p hsa-miR-365a-3p hsa-miR-374a-5p
1        5.373546        6.510108        4.570007        5.753530
2        6.183643        5.835624        7.709121        5.680214
3        5.164371        6.420695        7.435070        7.362644
4        7.595281        5.599753        5.289629        4.772117
5        6.329508        4.629792        5.934932        5.488781
6        5.179532        6.987838        4.240531        5.268805
  hsa-miR-423-5p hsa-miR-590-5p OS OS_STAT IPI_AGE IPI_LDH IPI_ECOG
1       8.236323       6.881278 12       1       1       0        1
2       6.302265       6.742082  0       0       1       1        0
3       4.957493       6.147573  3       1       0       1        0
4       5.016458       6.485389  2       1       0       1        1
5       8.005719       6.151856  4       0       1       1        0
6       3.929429       6.041999  2       0       0       1        0
  IPI_STAGE IPI_EXBM
1         1        0
2         1        0
3         0        1
4         1        1
5         0        0
6         0        1
}
}
