% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data__datasets.R
\name{dummy_reinders_2020_training_dataset}
\alias{dummy_reinders_2020_training_dataset}
\title{Dummy Dataset}
\format{
See section Examples.
}
\source{
\url{https://doi.org/10.1080/10428194.2020.1861268}
}
\description{
This is \strong{not} the training dataset used in \emph{Reinders, Jörg, et
al. "Platform independent protein-based cell-of-origin subtyping of diffuse
large B-cell lymphoma in formalin-fixed paraffin-embedded tissue."
Scientific Reports 10.1 (2020): 1-11.}. URL:
\url{https://doi.org/10.1038/s41598-020-64212-z}.
Instead, this dataset was \strong{simulated completely from scratch} to allow for
a test of the imlui application without having access to the real dataset.
}
\examples{
\dontrun{
head(dummy_reinders_2020_training_dataset)
#   sp|P04233|HG2A_HUMAN sp|Q15063|POSTN_HUMAN sp|P51884|LUM_HUMAN
# 1             18.74709              20.79621            18.75927
# 2             20.36729              18.77595            20.08423
# 3             18.32874              20.68224            18.17816
# 4             23.19056              17.74127            20.31606
# 5             20.65902              22.86605            18.69083
# 6             18.35906              23.96080            23.53457
#   sp|Q9NS69|TOM22_HUMAN sp|P62841|RPS15_HUMAN sp|P16070|CD44_HUMAN
# 1              20.90037              20.81880             20.27244
# 2              19.96288              23.37775             20.81434
# 3              19.36386              23.17318             19.86069
# 4              18.14128              19.33818             19.50467
# 5              17.02508              15.42953             21.39110
# 6              17.84962              24.99532             22.29246
#   sp|P01871|IGHM_HUMAN sp|P18031|PTN1_HUMAN Intercept  OS OS_STAT IPI_AGE
# 1             21.78735             22.94776  22.14888 113       1       0
# 2             17.90540             21.35454  23.79131  22       0       0
# 3             23.94267             20.75993  18.79401  32       1       0
# 4             19.23274             19.61440  19.21826   9       1       0
# 5             23.30829             23.15578  19.16756   8       0       1
# 6             23.02443             21.19247  19.24869   1       0       0
#   IPI_LDH IPI_ECOG IPI_STAGE IPI_EXBM
# 1       0        0         0        0
# 2       1        1         0        1
# 3       1        0         1        0
# 4       1        1         1        0
# 5       1        0         0        0
# 6       0        1         1        0
}
}
