% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data__datasets.R
\name{dummy_staiger_2020_test1_dataset}
\alias{dummy_staiger_2020_test1_dataset}
\title{Dummy Dataset}
\format{
A data.frame with 181 obs. of 19 variables. For details see section
Examples.
}
\source{
\url{https://www.nature.com/articles/s41375-019-0573-y/}
}
\description{
The LAMIS signature as described in \emph{Staiger, A.M.,
Altenbuchinger, M., Ziepert, M. et al. A novel lymphoma-associated
macrophage interaction signature (LAMIS) provides robust risk
prognostication in diffuse large B-cell lymphoma clinical trial cohorts of
the DSHNHL. Leukemia 34, 543-552 (2020)} was tested on two different test
sets, called validation set I and validation set II. This dataset is
\strong{neither} of those two datasets, but instead \strong{simulated completely from
scratch} to allow for a test of the imlui application without having access
to the real datasets. This dataset is meant as dummy replacement for the real
\emph{Validation set I} and has the amount of samples (181) as described in the
publication.
}
\examples{
\dontrun{
head(dummy_staiger_2020_test1_dataset)
#     BCL2A1      BSG     CCL5     CCND2    CD163     CD47    CPNE3    CPT1A
# 1 11.37355 12.98390 11.36446 12.018607 11.02445 10.96710 13.26201 11.49082
# 2 12.18364 12.21992 11.57002 13.318449 12.07066 12.65920 11.59423 11.14299
# 3 11.16437 10.53275 11.83068 11.934168 10.48140 12.23783 12.66676 13.61617
# 4 13.59528 12.52102 12.61222 11.299704 12.86378 12.71528 12.16464 12.99387
# 5 12.32951 11.84125 12.67834 12.537326 12.50157 11.06152 13.78152 12.69685
# 6 11.17953 13.46459 12.56795  9.798218 11.64522 12.09535 12.71121 13.69997
#       CSF1    CXCR4   DPYSL3      FYN     IL16      MME      MPST    PSAT1
# 1 12.23356 11.98254 11.92833 12.84547 11.45736 11.89782 12.805547 12.11641
# 2 11.73096 11.75903 11.86686 12.46725 13.97869 12.16720 11.787799 11.53369
# 3 11.34145 10.20801 11.91192 11.59794 10.63389 12.45661 11.147571 13.84623
# 4 12.06335 14.04969 12.91779 12.92353 12.42490 11.49843 10.703325 10.70860
# 5 12.05448 10.88278 12.03144 11.99194 11.02369 12.22694  9.455339 12.96074
# 6 12.24701 10.64645 13.35893 13.03378 13.19297 11.29137 12.536258 12.23254
#        TBP  OS OS_STAT IPI_AGE IPI_LDH IPI_ECOG IPI_STAGE IPI_EXBM
# 1 12.44005 120       1       0       0        0         1        0
# 2 10.47804  51       0       0       0        0         0        0
# 3 12.10962  17       1       0       1        1         0        0
# 4 12.18904   2       1       0       1        0         0        0
# 5 12.78104  12       0       1       0        0         1        0
# 6 11.95349   4       0       0       1        0         0        1
}
}
