% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fep.R
\name{FEP}
\alias{FEP}
\title{Visualize numeric predictions \code{ys} made for datasets \code{names(ys)}}
\usage{
FEP(predictions, density_lines = TRUE, rug = TRUE, legend = TRUE)
}
\arguments{
\item{predictions}{as the name says (named list of named numeric vectors)}

\item{density_lines}{add density lines to the plot (logical)? (logical)}

\item{rug}{add rugs to the plot? (logical)}

\item{legend}{add a legend to the plot explaining \code{names(predictions)}? (logical)}

\item{histogram}{add a histogram to the plot? (logical)}
}
\value{

}
\description{
Visualize numeric predictions \code{ys} made for datasets \code{names(ys)}
}
\examples{
TODO
}
