% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/740_pkg_utils.R
\name{getdata}
\alias{getdata}
\title{Get Symbol from Package}
\usage{
getdata(sym, typ, pkg, transpose)
}
\arguments{
\item{sym}{Symbol to get as character.}

\item{typ}{Type of object (either "dataset" for dataset, or "model" for model).}

\item{pkg}{In which package to look for \code{sym}.}

\item{transpose}{Transpose object after initialisation? (only used if \code{typ} == "dataset")}
}
\value{
\code{obj: get(sym)}
}
\description{
Like normal get but additionally calls \code{data} if the object does not exist yet and performs some
initial transformation to make the data compliant to the format expected by imlui functions.
}
