% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/070_hndl_oauth_github.R
\name{hndl_auth_github}
\alias{hndl_auth_github}
\title{Set auth state based on Github OAuth Token}
\usage{
hndl_auth_github(ses)
}
\arguments{
\item{ses}{List of session specific objects as returned by \code{init_data} once
for every user session}
}
\description{
Set auth state based on Github OAuth Token
}
\details{
Should be called once at the beginning of the server function
}
