% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html.R
\name{html__inp_models}
\alias{html__inp_models}
\title{Dropdown menu to select Models}
\usage{
html__inp_models(ses, id, n = 1, multiple = TRUE, dataset_inp_id = NULL)
}
\arguments{
\item{ses}{session environment}

\item{id}{id for generated html element}

\item{n}{maximum number of allowed selections}

\item{multiple}{Set TRUE if multiple selections are allowed}

\item{dataset_inp_id}{ID of a \code{html__inp_model} element. If not NULL, the
set of models availble in the returned HTML pickerInput will be restricted to
models compatible to the selected datasets from \code{dataset_inp_id}.}
}
\description{
Dropdown menu to select Models.
}
\details{
Documentation Links:
\itemize{
\item \url{https://dreamrs.github.io/shinyWidgets/reference/pickerInput.html}
\item \url{https://dreamrs.github.io/shinyWidgets/reference/pickerOptions.html}
\item \url{https://developer.snapappointments.com/bootstrap-select/options/}
}

Unfortunately, the corresponding bootstrap widget does not allow to specify
multiple selections together with a minimum number of selections. Therefor we
only have parameter n, for number of max selections. However, if multiple
is set to TRUE, exactly one selection is required.
}
