% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/html__navbar.R
\name{html__navbar__main}
\alias{html__navbar__main}
\title{Main Navbar Element}
\usage{
html__navbar__main(id, brand, ..., collapsible = TRUE)
}
\arguments{
\item{id}{e.g. "navbar-1"}

\item{brand}{e.g. tags$span(class = "navbar-brand", "IMLUI")}

\item{...}{ul elements with class navbar-left or navbar-right, as generated
by \code{html__navbar__left()} or \code{html__navbar__right()}}

\item{collapsible}{whether navbar should be collapsible or not}
}
\description{
Main Bootstrap navbar element
}
