% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/045_init_dataset_rs.R
\name{init_dataset_reactives}
\alias{init_dataset_reactives}
\title{Return list of reactives related to Dataset Metadata}
\usage{
init_dataset_reactives(input, rv)
}
\arguments{
\item{input}{Input object as handed over to server functions by shiny}

\item{rv}{List of reactive values as returned by
\code{\link[=init_reactive_values]{init_reactive_values()}}}

\item{db}{DB object as returned by \code{DB$new()}}
}
\value{
list with following reactive elements:
tbl, ids, symbols, displaynames, pkgs, transpose
reactives returning vectors of all ids/symbols/displaynames/...
accessible by the current user
df, df_cat, df_num, features, samples
lists of reactives (one reactive per dataset)
symbols_list, df_list, df_cat_list, df_num_list, samples_list, features_list
reactives returning vectors of ids/symbols/displaynames/...
selected by the current user
}
\description{
Return list of reactives related to Dataset Metadata
}
