% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/132_init_model_reactives.R
\name{init_model_reactives}
\alias{init_model_reactives}
\title{Return list of reactives related to Model Metadata}
\usage{
init_model_reactives(rvs, db)
}
\arguments{
\item{rvs}{List of reactive values as returned by
\code{\link[=init_reactive_values]{init_reactive_values()}}}

\item{db}{DB object as returned by \code{DB$new()}}
}
\value{
List of reactives with following names: df, ids,
}
\description{
Return list of reactives related to Model Metadata
}
