% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/050_init_rs_model.R
\name{init_rs_model}
\alias{init_rs_model}
\title{Return list of reactives related to Model Metadata}
\usage{
init_rs_model(input, rv)
}
\arguments{
\item{input}{Input object as handed over to server functions by shiny}

\item{rv}{List of reactive values as returned by
\code{\link[=init_reactive_values]{init_reactive_values()}}}
}
\value{
list with following elements:
\itemize{
\item \code{ids}: reactive, all elements of \code{tbl$ID} that may be accessed by the
current user
\item \code{symbols}: reactive, \code{R} objects corresponding to \code{ids()}
\item \code{displaynames}: reactive, display names corresponding to \code{ids()}
\item \code{pkgs}: reactive, R packages providing \code{symbols}
\item \code{features}: list of reactives, feature names for each symbol in \code{symbols()}
\item \code{betas}: list of reactives, model betas for each symbol in \code{symbols()}
\item \code{symbols_list}: reactive, list of symbols currently selected by the user
\item \code{features_list}: reactive, list of feature name vectors corresponding to
\code{symbols_list()}
\item \code{betas_list}: reactive, list of  model param vectors corresponding to
\code{symbols_list()}
}
}
\description{
Return list of reactives related to Model Metadata
}
