% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data__models.R
\name{ipi}
\alias{ipi}
\title{The International Prognostic Index (IPI)}
\format{
A named vector of length 5. See Examples for details
}
\source{
\url{https://doi.org/10.1056/NEJM199309303291402}
}
\description{
The International Prognostic Index (IPI) as described in
\emph{A predictive model for aggressive non-Hodgkin's lymphoma. The International
Non-Hodgkin's Lymphoma Prognostic Factors Project. N Engl J Med
1993;329(14):987-94}.

The IPI works by assigning one point for each the following factors:
\itemize{
\item Age greater than 60 years
\item Stage III or IV disease
\item Elevated serum LDH
\item ECOG/Zubrod performance status of 2, 3, or 4
\item More than 1 extranodal site
}

The sum of points can be used to separate samples into following risk groups:
\itemize{
\item Low risk (0-1 points)
\item Low-intermediate risk (2 points)
\item High-intermediate risk (3 points)
\item High risk (4-5 points)
}
}
\examples{
\dontrun{
as.data.frame(ipi)
#                   ipi
# age_gt_60           1  # Age greater than 60 years
# stage_3_4           1  # Stage III or IV disease
# lhd_elevated        1  # Elevated serum LDH
# ecog_2_3_4          1  # ECOG/Zubrod performance status of 2, 3, or 4
# ex_nod_sites_gt_1   1  # More than 1 extranodal site
}
}
