% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/710_pkg_datasets.R
\docType{data}
\name{lamis_signature}
\alias{lamis_signature}
\title{The LAMIS signature}
\format{
A named vector of length 17.
}
\source{
\url{https://www.nature.com/articles/s41375-019-0573-y/}
}
\usage{
lamis_signature
}
\description{
The LAMIS signature as described in \emph{Staiger, A.M., Altenbuchinger, M.,
Ziepert, M. et al. A novel lymphoma-associated macrophage interaction
signature (LAMIS) provides robust risk prognostication in diffuse large
B-cell lymphoma clinical trial cohorts of the DSHNHL. Leukemia 34, 543-552
(2020). https://doi.org/10.1038/s41375-019-0573-y}
}
\examples{
round(lamis_signature, 3)
#   CSF1  CPT1A  CD163   CCL5 BCL2A1  CXCR4    MME    BSG   CD47   IL16
# -0.685  0.685  0.155 -0.146 -0.134  0.122 -0.089  0.089  0.079 -0.065
#    FYN   MPST DPYSL3  CCND2  CPNE3    TBP  PSAT1
# -0.058  0.050 -0.049  0.040 -0.031  0.031  0.005
}
\keyword{datasets}
