% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/71_pkg_datasets.R
\docType{data}
\name{lamis_test1}
\alias{lamis_test1}
\title{The LAMIS test data 1 (validaton set I).}
\format{
A data.frame with 181 obs. of 20524 variables. Columns 1-22
contain the following variables:\preformatted{1. Gender [Levels: female, male]
2. Age [Range: 14-88]
3. Tissue [Levels: lymphnode]
4. Disease [Levels: Lymphoma]
5. Individual [Range: 304-1061]
6. diagnosis [Levels: DLBCL]
7. ABCGCB [Levels: ABCDLBCL, GCBDLBCL, UnclassifiedDLBCL]
8. Follow.up.status [Levels: ALIVE, DEAD]
9. Follow.up.years [Range: 0-21.78]
10. Chemotherapy [Levels: CHOP-LikeRegimen]
11. ECOG.performance.status [Levels: 0, 1, 2, 3, 4]
12. Stage [Levels: 1, 2, 3, 4]
13. LDH.ratio [Range: NA-NA]
14. extranodal.sites [Levels: 0, 1]
15. IPI.scores [Levels: 0, 1, 2, 3, 4]
16. IPI.rating [Levels: 0, 1, 2]
17. ABCGCB.score [Levels: 0, 0.52, 1]
18. IPI_AGE [Levels: 0, 1]
19. IPI_LDH [Levels: 0, 0.51, 1]
20. IPI_ECOG [Levels: 0, 0.24, 1]
21. IPI_STAGE [Levels: 0, 0.39, 1]
22. IPI_EXBM [Levels: 0, 0.08]
}

#' Columns 23-20524 contain log2 transformed gene expression values.
}
\source{
\url{https://www.nature.com/articles/s41375-019-0573-y/}
}
\usage{
lamis_test1
}
\description{
The LAMIS test data as described in \emph{Staiger, A.M., Altenbuchinger, M.,
Ziepert, M. et al. A novel lymphoma-associated macrophage interaction
signature (LAMIS) provides robust risk prognostication in diffuse large
B-cell lymphoma clinical trial cohorts of the DSHNHL. Leukemia 34, 543-552
(2020). https://doi.org/10.1038/s41375-019-0573-y}. Called \emph{validation set
I} in the paper.
}
\examples{
lamis_test1[1:3, 1:25]
#           Gender Age    Tissue  Disease Individual diagnosis   ABCGCB
# GSM274895   male  52 lymphnode Lymphoma        549     DLBCL GCBDLBCL
# GSM274896   male  75 lymphnode Lymphoma        553     DLBCL GCBDLBCL
# GSM274897 female  76 lymphnode Lymphoma        689     DLBCL ABCDLBCL
#           Follow.up.status Follow.up.years     Chemotherapy
# GSM274895             DEAD            2.68 CHOP-LikeRegimen
# GSM274896             DEAD            0.82 CHOP-LikeRegimen
# GSM274897             DEAD            2.54 CHOP-LikeRegimen
#           ECOG.performance.status Stage LDH.ratio extranodal.sites
# GSM274895                       1     3        NA                0
# GSM274896                       0     3      1.12                0
# GSM274897                       3     1      1.01                0
#           IPI.scores IPI.rating ABCGCB.score IPI_AGE   IPI_LDH IPI_ECOG
# GSM274895         NA         NA            1       0 0.5071225        0
# GSM274896          3          1            1       1 1.0000000        0
# GSM274897          3          1            0       1 1.0000000        1
#           IPI_STAGE IPI_EXBM     RFC2    HSPA6     PAX8
# GSM274895 0.3915858        0 7.080452 6.478552 4.987939
# GSM274896 0.3915858        0 7.924949 6.470217 5.111094
# GSM274897 0.0000000        0 7.188202 6.206049 5.114223
}
\keyword{datasets}
