% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{lamis_test2}
\alias{lamis_test2}
\title{The LAMIS test data 2 (validaton set II).}
\format{
A data.frame with 466 obs. of 168 variables. Columns 1-23 contain
the following variables:\preformatted{
1. STUDIDn: study identification number [7:RICOVER-60, 70:RICOVER-noRTh,
   9:MegaCHOEP Phase-III, 90:MegaCHOEP observation Phase-II, 12:DENSER,
   120:SMARTER, 999:MInT]
2. PATSTUID: patient identification number within trial [Levels: 1, 100,
   101, 1013, ..., D-70-18, D-70-22, D-70-25, D-70-3, D-70-8]
3. ARMID: TODO [Levels: 1, 3, 4, 5, 22]
4. GENDER: gender [0:male, 1:female]
5. AGE: age in years
6. IPI_AGE: IPI factor age [0:age<=60, 1:age>60]
7. IPI_LDH: IPI factor LDH [0:LDH<=N, 1:LDH>N]
8. IPI_ECOG: IPI factor ECOG>1 [0:ECOG<=1, 1:ECOG>1]
9. IPI_STAGE: IPI factor stage III/IV [0=stage I/II, 1=stage III/IV]
10. IPI_EXBM: IPI factor extralymphatic involvement (EL) [0:EL<=1, 1:EL>1]
11. IPI: IPI score [0-5]
12. AA_IPI: age-adjusted (< 60 years) IPI score [0-3]
13. EXBM: extralymphatic involvement [0=no, 1=yes]
14. BULK: bulk involvement >7.5cm (>5.0cm for STUDIDn=999) [0=no, 1=yes]
15. BSYMP: B symptoms [0=no, 1=yes]
16. BM: bone marrow involvement [0=no, 1=yes]
17. REFPATHO: reference pathology TODO
18. EFS: event-free survival in months
19. EFS_STAT: event-free survival status [0=censored, 1=event]
20. PFS: progression-free survival in months
21. PFS_STAT: progression-free survival status [0=censored, 1=event]
22. OS: overall survival in months
23. OS_STAT: overall survival status [0=censored 1=death from any cause]
}

Columns 24-20524 contain log2 transformed gene expression values.
}
\source{
\url{https://www.nature.com/articles/s41375-019-0573-y/}
}
\usage{
lamis_test2
}
\description{
The LAMIS test data as described in \emph{Staiger, A.M., Altenbuchinger, M.,
Ziepert, M. et al. A novel lymphoma-associated macrophage interaction
signature (LAMIS) provides robust risk prognostication in diffuse large
B-cell lymphoma clinical trial cohorts of the DSHNHL. Leukemia 34, 543-552
(2020). https://doi.org/10.1038/s41375-019-0573-y}. Called \emph{validation set
II} in the paper.
}
\examples{
imlui::lamis_test2[1:3, 1:25]
#                STUDIDn PATSTUID ARMID DOSELEVEL GENDER AGE IPI_AGE
# Ricover60.S99        7       99     4        NA      1  64       1
# Ricover60.S101       7      101     3        NA      0  67       1
# Ricover60.S102       7      102     3        NA      0  73       1
#                IPI_LDH IPI_ECOG IPI_STAGE IPI_EXBM IPI AA_IPI EXBM BULK
# Ricover60.S99        0        0         1        1   3     NA    1    0
# Ricover60.S101       1        0         1        0   3     NA    0    1
# Ricover60.S102       0        0         0        0   1     NA    1    0
#                BSYMP BM REFPATHO      EFS EFS_STAT      PFS PFS_STAT
# Ricover60.S99      0  1     1040 109.9630        0 109.9630        0
# Ricover60.S101     1  0     1010   4.5339        1   4.5339        1
# Ricover60.S102     0  0     1040  54.5051        1  54.5051        1
#                      OS OS_STAT    BCL2A1
# Ricover60.S99  109.9630       0  9.601771
# Ricover60.S101  10.5133       1 10.251482
# Ricover60.S102  54.5051       1  7.714246
}
\keyword{datasets}
