% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/71_pkg_datasets.R
\docType{data}
\name{lamis_train}
\alias{lamis_train}
\title{The LAMIS training data}
\format{
A data.frame with 233 obs. of 20524 variables. Columns 1-22
contain the following variables:\preformatted{1. Gender [Levels: female, male]
2. Age [Range: 17-92]
3. Tissue [Levels: lymphnode]
4. Disease [Levels: Lymphoma]
5. Individual [Range: 2-2814]
6. diagnosis [Levels: DLBCL]
7. ABCGCB [Levels: ABCDLBCL, GCBDLBCL, UnclassifiedDLBCL]
8. Follow.up.status [Levels: ALIVE, DEAD]
9. Follow.up.years [Range: 0-10.29]
10. Chemotherapy [Levels: R-CHOP-LikeRegimen]
11. ECOG.performance.status [Levels: 0, 1, 2, 3, 4]
12. Stage [Levels: 1, 2, 3, 4]
13. LDH.ratio [Range: 0-30.76]
14. extranodal.sites [Levels: 0, 1, 2, 3, 4, 5]
15. IPI.scores [Levels: 0, 1, 2, 3, 4, 5]
16. IPI.rating [Levels: 0, 1, 2]
17. ABCGCB.score [Levels: 0, 0.52, 1]
18. IPI_AGE [Levels: 0, ]
19. IPI_LDH [Levels: 0, 0.50, 1]
20. IPI_ECOG [Levels: 0, 0.23, 1]
21. IPI_STAGE [Levels: 0, 0.39, 1]
22. IPI_EXBM [Levels: 0, 0.07, 1]
}

Columns 23-20524 contain log2 transformed gene expression values.
}
\source{
\url{https://www.nature.com/articles/s41375-019-0573-y/}
}
\usage{
lamis_train
}
\description{
The LAMIS training data as described in \emph{Staiger, A.M., Altenbuchinger, M.,
Ziepert, M. et al. A novel lymphoma-associated macrophage interaction
signature (LAMIS) provides robust risk prognostication in diffuse large
B-cell lymphoma clinical trial cohorts of the DSHNHL. Leukemia 34, 543-552
(2020). https://doi.org/10.1038/s41375-019-0573-y}
}
\examples{
lamis_train[1:3, 1:25]
#           Gender Age    Tissue  Disease Individual diagnosis   ABCGCB
# GSM275076   male  73 lymphnode Lymphoma       1179     DLBCL ABCDLBCL
# GSM275077 female  71 lymphnode Lymphoma       1181     DLBCL GCBDLBCL
# GSM275078   male  71 lymphnode Lymphoma       1184     DLBCL GCBDLBCL
#           Follow.up.status Follow.up.years       Chemotherapy
# GSM275076            ALIVE            2.52 R-CHOP-LikeRegimen
# GSM275077            ALIVE            1.71 R-CHOP-LikeRegimen
# GSM275078            ALIVE            4.74 R-CHOP-LikeRegimen
#           ECOG.performance.status Stage LDH.ratio extranodal.sites
# GSM275076                       0     4      0.92                1
# GSM275077                       1     2      0.63                0
# GSM275078                       0     3      0.74                0
#           IPI.scores IPI.rating ABCGCB.score IPI_AGE IPI_LDH IPI_ECOG
# GSM275076          2          1            0       1       0        0
# GSM275077          1          0            1       1       0        0
# GSM275078          2          1            1       1       0        0
#           IPI_STAGE IPI_EXBM     RFC2    HSPA6     PAX8
# GSM275076 1.0000000        0 8.059630 5.399721 4.788501
# GSM275077 0.0000000        0 8.578449 6.115090 4.977569
# GSM275078 0.3915858        0 7.934017 5.191435 5.064762
}
\keyword{datasets}
