% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dyn.R
\name{mock_input_obj}
\alias{mock_input_obj}
\title{Return a shiny server input object (for Testing Purposes)}
\usage{
mock_input_obj(...)
}
\arguments{
\item{...}{Properties of returned input object}
}
\description{
Return a shiny server input object (for Testing Purposes)
}
\examples{
local({
  rc_enabled <- getOption("shiny.suppressMissingContextError", FALSE)
  shiny::reactiveConsole(TRUE)
  on.exit(shiny::reactiveConsole(rc_enabled), add = TRUE)
  input <- imlui:::mock_input_obj(x = 1, y = 2)
  input$x == 1 && input$y == 2
})
}
