% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_mocks.R
\name{mock_rv_obj}
\alias{mock_rv_obj}
\title{Return a reactivevalues Object (for Testing Purposes)}
\usage{
mock_rv_obj(user_id = NULL)
}
\arguments{
\item{user_id}{Either 'toscm', 'max', 'public' or NULL}
}
\description{
Return a reactivevalues Object (for Testing Purposes)
}
