% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_mocks.R
\name{mock_session_obj}
\alias{mock_session_obj}
\title{Return a MockShinySession object (for Testing Purposes)}
\usage{
mock_session_obj()
}
\value{
R6 MockShinySession object
}
\description{
Return a MockShinySession object (for Testing Purposes)
}
