% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{nl_nv_2_df}
\alias{nl_nv_2_df}
\title{Convert named list of named vectors \code{nlnv} to dataframe \code{df} with:\preformatted{df[, 1] == names(nlnv)
df[, 2] == names(nlnv[i])
df[, 3] == nlnv[[i]][j]

[, 1]: R:,\%201
[, 2]: R:,\%202
[i]: R:i
[, 3]: R:,\%203
[i]: R:i
}}
\usage{
nl_nv_2_df(nl_nv, col1 = "dataset", col2 = "sample", col3 = "y")
}
\arguments{
\item{x}{Numeric value}
}
\value{

}
\description{
Convert named list of named vectors \code{nlnv} to dataframe \code{df} with:\preformatted{df[, 1] == names(nlnv)
df[, 2] == names(nlnv[i])
df[, 3] == nlnv[[i]][j]

[, 1]: R:,\%201
[, 2]: R:,\%202
[i]: R:i
[, 3]: R:,\%203
[i]: R:i
}
}
\examples{
# nl_nv = List of 3:
# $ lamis_train: Named num [1:233] GSM275076 = 0.6801, GSM275077 = 1.0086 ...
# $ lamis_test1: Named num [1:181] GSM274895 = 0.505, GSM274896 = 1.541 ...
# $ lamis_test2: Named num [1:466] Ricover60.S99 = 0.822, Ricover60.S101 = 1.203 ...

head(nl_nv_2_df(nl_nv))
#       dataset   samples           y
# 1 lamis_train GSM275076 -0.68014232
# 2 lamis_train GSM275077 -1.00861666
# 3 lamis_train GSM275078 -1.37423898
# 4 lamis_train GSM275079 -0.79874698
# 5 lamis_train GSM275080  0.03045401
# 6 lamis_train GSM275081 -0.61785528

str(nl_nv_2_df(nl_nv))
# 'data.frame':   880 obs. of  3 variables:
#  $ dataset: chr  "lamis_train" "lamis_train" "lamis_train" "lamis_train" ...
#  $ samples: chr  "GSM275076" "GSM275077" "GSM275078" "GSM275079" ...
#  $ y      : num  -0.6801 -1.0086 -1.3742 -0.7987 0.0305 ...
}
