% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data__models.R
\name{nordmo_2020_pfs_signature}
\alias{nordmo_2020_pfs_signature}
\title{miRNA Progression Free Survival signature for DLBCL patients}
\format{
A named vector of length 6.
}
\source{
\url{https://doi.org/10.1080/10428194.2020.1861268}
}
\description{
The progression free survival signature described in Nordmo,
Carmen, Gunther Glehr, Michael Altenbuchinger, Rainer Spang, Marita Ziepert,
Heike Horn, Annette M. Staiger, et al. "Identification of a MiRNA Based
Model to Detect Prognostic Subgroups in Patients with Aggressive B-Cell
Lymphoma". Leukemia & Lymphoma 62, no. 5 (16 April 2021): 1107-15.
https://doi.org/10.1080/10428194.2020.1861268.
}
\examples{
nordmo_pfs_signature = c(
    "miR-106b-5p" = +0.967200,
    "miR-130a-3p" = -0.153873,
    "miR-365a-3p" = -0.945600,
    "miR-374a-5p" = +0.098353,
    "miR-423-5p"  = -0.138210,
    "miR-590-5p"  = +0.172130
)
}
