% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nordmo_os_signature}
\alias{nordmo_os_signature}
\title{miRNA Overall Survival signature for DLBCL patients}
\format{
A named vector of length 6.
}
\source{
\url{https://doi.org/10.1080/10428194.2020.1861268}
}
\usage{
nordmo_os_signature
}
\description{
The overall survival signature described in Nordmo, Carmen, Gunther Glehr,
Michael Altenbuchinger, Rainer Spang, Marita Ziepert, Heike Horn, Annette M.
Staiger, et al. "Identification of a MiRNA Based Model to Detect Prognostic
Subgroups in Patients with Aggressive B-Cell Lymphoma". Leukemia & Lymphoma
62, no. 5 (16 April 2021): 1107-15.
https://doi.org/10.1080/10428194.2020.1861268.
}
\examples{
nordmo_os_signature = c(
    "hsa_miR-106b-5p" = +0.189710,
    "hsa_miR-130a-3p" = -0.233681,
    "hsa_miR-186-5p"  = +0.069890,
    "hsa_miR-374b-5p" = +0.003091,
    "hsa_miR-423-5p"  = -0.069890,
    "hsa_miR-590-5p"  = +0.040880
)
}
\keyword{datasets}
