% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{nordmo_pfs_signature}
\alias{nordmo_pfs_signature}
\title{miRNA Progression Free Survival signature for DLBCL patients}
\format{
A named vector of length 6.

A 209 x 800 Dataframe\preformatted{       hsa-miR-892a hsa-miR-128 ... hsa-miR-454-3p hsa-miR-4485
S14901     4.087463    6.022368 ...       5.247928     4.523562
S15201     5.129283    5.357552 ...       6.129283     4.643856
   ...          ...         ... ...            ...          ...
S18101     4.087463    4.807355 ...       5.426265     3.807355
S18201     3.459432    4.807355 ...       5.000000     4.169925
}
}
\source{
\url{https://doi.org/10.1080/10428194.2020.1861268}

\url{https://doi.org/10.1080/10428194.2020.1861268}
}
\usage{
nordmo_pfs_signature

nordmo_pfs_signature
}
\description{
The progression free survival signature described in Nordmo, Carmen, Gunther
Glehr, Michael Altenbuchinger, Rainer Spang, Marita Ziepert, Heike Horn,
Annette M. Staiger, et al. "Identification of a MiRNA Based Model to Detect
Prognostic Subgroups in Patients with Aggressive B-Cell Lymphoma". Leukemia
& Lymphoma 62, no. 5 (16 April 2021): 1107-15.
https://doi.org/10.1080/10428194.2020.1861268.

The (already log2 transformed) training data used in Nordmo, Carmen, Gunther
Glehr, Michael Altenbuchinger, Rainer Spang, Marita Ziepert, Heike Horn,
Annette M. Staiger, et al. "Identification of a MiRNA Based Model to Detect
Prognostic Subgroups in Patients with Aggressive B-Cell Lymphoma". Leukemia
& Lymphoma 62, no. 5 (16 April 2021): 1107-15.
https://doi.org/10.1080/10428194.2020.1861268.
}
\examples{
nordmo_pfs_signature = c(
    "miR-106b-5p" = +0.967200,
    "miR-130a-3p" = -0.153873,
    "miR-365a-3p" = -0.945600,
    "miR-374a-5p" = +0.098353,
    "miR-423-5p"  = -0.138210,
    "miR-590-5p"  = +0.172130
)
}
\keyword{datasets}
