% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/920_patched_cached_func_with_file.R
\name{patched_cached_func_with_file}
\alias{patched_cached_func_with_file}
\title{Patched version of \code{shiny::cachedFuncWithFile}}
\usage{
patched_cached_func_with_file(dir, file, func, case.sensitive = FALSE)
}
\arguments{
\item{dir}{required by original func}

\item{file}{required by original func}

\item{func}{required by original func}

\item{case.sensitive}{required by original func}
}
\description{
Patched version of \verb{shiny::cachedFuncWithFile. Code changes are inspired by function }shiny::initAutoReloadMonitor\verb{from file "shinyapp.R". See [README.md](README.md#shiny-bugfixworkaround-autoreload-for-packages) for details. Used by}dyn_serve_app_in_dev_mode`.
}
