% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot__fep}
\alias{plot__fep}
\title{Visualize Feature Effects}
\usage{
plot__fep(
  b,
  X,
  s,
  debug = FALSE,
  n_extreme = 10,
  threshold = 0.991,
  show_score = FALSE
)
}
\arguments{
\item{b}{betas of linear model (named m*1 vector)}

\item{X}{dataset (n*m matrix)}

\item{s}{sample of interest}

\item{debug}{draw colored borders to debug the plot}

\item{n_extreme}{how many samples from X to draw into the plot?
Samples are taken from both end of the y spectrum, i.e.}

\item{threshold}{threshold for classification setting}

\item{show_score}{show score in plot}
}
\value{
`par(opar)``
}
\description{
Visualize feature effects for linear models of the form  y = X
\itemize{
\item b
}
}
