% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot__msd}
\alias{plot__msd}
\title{Mean Standard Deviation Plot}
\usage{
plot__msd(dd, xx, xt, yt, ff = NULL, shape = ".")
}
\arguments{
\item{dd}{list(char), dataset names}

\item{xx}{list(data.frame), covariate dataframes (must contain only numeric
columns)}

\item{xt}{char, how to transform the xaxis for plotting (options are
'identity', 'log2', 'log10' and 'rank')}

\item{yt}{char, how to transform the xaxis for plotting (options are
'identity', 'log2', 'log10' and 'rank')}

\item{ff}{vector(char), feature names to draw, if \code{util__is_none(ff)}, all
features are drawn}

\item{shape}{char, character shape to use, not implemented yet}
}
\value{
TODO
}
\description{
The mean against standard deviation for each feature in a list
of datasets. The datasets must not contain non numeric columns.
}
