% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fep.r
\name{plot_feature_contribution}
\alias{plot_feature_contribution}
\title{Visualize feature contributions for linear models of the form  y = X * b}
\usage{
plot_feature_contribution()
}
\value{

}
\description{
Visualize feature contributions for linear models of the form  y = X * b
}
\examples{
TODO
}
