% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fep.R
\name{plot_feature_effects}
\alias{plot_feature_effects}
\title{Visualize feature effetcs for linear models of the form  y = X * b}
\usage{
plot_feature_effects(b, X, s, debug = FALSE, n_extreme = 10, threshold = 0.991)
}
\arguments{
\item{b}{betas of linear model (named m*1 vector)}

\item{X}{dataset (n*m matrix)}

\item{s}{sample of interest}

\item{debug}{draw colored borders to debug the plot}

\item{n_extreme}{how many samples from X to draw into the plot?
Samples are taken from both end of the y spectrum, i.e.}

\item{threshold}{threshold for classification setting}
}
\value{

}
\description{
Visualize feature effetcs for linear models of the form  y = X * b
}
\examples{
TODO
}
