% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data__models.R
\name{reinders_2020_coo_signature}
\alias{reinders_2020_coo_signature}
\title{Protein COO signature for DLBCL patients}
\format{
A named vector of length 9. See example for the full vector.
}
\source{
\url{https://doi.org/10.1080/10428194.2020.1861268}
}
\description{
The COO signature described in \emph{Reinders, Jörg, et al.
"Platform independent protein-based cell-of-origin subtyping of diffuse
large B-cell lymphoma in formalin-fixed paraffin-embedded tissue."
Scientific Reports 10.1 (2020): 1-11.}. For details see
\url{https://doi.org/10.1038/s41598-020-64212-z}.
The exact signature beta values were copied from \href{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC7217957/table/Tab1/?report=objectonly}{Table 1}
of the paper.
}
\examples{
reinders_2020_coo_signature <- c(
    "sp|P04233|HG2A_HUMAN" = 4.489047570,
    "sp|Q15063|POSTN_HUMAN" = 1.551003753,
    "sp|P51884|LUM_HUMAN" = 1.490571730,
    "sp|Q9NS69|TOM22_HUMAN" = 1.019845999,
    "sp|P62841|RPS15_HUMAN" = -0.888995137,
    "sp|P16070|CD44_HUMAN" = -1.624465232,
    "sp|P01871|IGHM_HUMAN" = -2.556118873,
    "sp|P18031|PTN1_HUMAN" = -3.480889810,
    "Intercept" = -1.25855429
)
}
