% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{serve_web_app}
\alias{serve_web_app}
\title{Serve the IMLUI Web App}
\usage{
serve_web_app(
  config_file = NULL,
  config_dir = NULL,
  data_dir = NULL,
  port = PORT,
  host = "0.0.0.0",
  suppress_log_messages = FALSE,
  suppress_debug_messages = FALSE,
  devmode = util__in_devmode(),
  autoreload = TRUE,
  via_shinyserver = FALSE,
  launch.browser = FALSE,
  quiet = TRUE,
  display.mode = "normal",
  test.mode = FALSE
)
}
\arguments{
\item{config_file}{TODO}

\item{config_dir}{TODO}

\item{data_dir}{TODO}

\item{port}{TODO}

\item{host}{TODO}

\item{suppress_log_messages}{TODO}

\item{suppress_debug_messages}{TODO}

\item{devmode}{TODO}

\item{autoreload}{TODO}

\item{via_shinyserver}{TODO}

\item{launch.browser}{TODO}

\item{quiet}{TODO}

\item{display.mode}{TODO}

\item{test.mode}{TODO}
}
\value{
NULL or object of class \code{shiny.appobj}.
}
\description{
Serves the IMLUI Web App via HTTP or returns an
object of class \code{shiny.appobj} if argument \code{via_shinyserver} is true.
}
