% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data__models.R
\name{staiger_2020_survival_signature_lamis}
\alias{staiger_2020_survival_signature_lamis}
\title{The LAMIS signature}
\format{
A named vector of length 17.
}
\source{
\url{https://www.nature.com/articles/s41375-019-0573-y/}
}
\description{
The LAMIS signature as described in \emph{Staiger, A.M.,
Altenbuchinger, M., Ziepert, M. et al. A novel lymphoma-associated
macrophage interaction signature (LAMIS) provides robust risk
prognostication in diffuse large B-cell lymphoma clinical trial cohorts of
the DSHNHL. Leukemia 34, 543-552 (2020)}. For details see
\url{https://doi.org/10.1038/s41375-019-0573-y}.
The beta values were taken for Supplemnatry \href{https://static-content.springer.com/esm/art\%3A10.1038\%2Fs41375-019-0573-y/MediaObjects/41375_2019_573_MOESM3_ESM.docx}{Table S2}
of the publication.
}
\examples{
\dontrun{
round(staiger_2020_survival_signature_lamis, 3)
# BCL2A1    BSG   CCL5  CCND2  CD163   CD47  CPNE3  CPT1A   CSF1  CXCR4 DPYSL3
# -0.134  0.089 -0.146  0.040  0.155  0.079 -0.031  0.685 -0.685  0.122 -0.049
#    FYN   IL16    MME   MPST  PSAT1    TBP
# -0.058 -0.065 -0.089  0.050  0.005  0.031
}
}
