% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{web_app}
\alias{web_app}
\title{IMLUI Web App}
\usage{
web_app(
  config_file = NULL,
  config_dir = NULL,
  host = "0.0.0.0",
  launch.browser = FALSE,
  port = PORT,
  quiet = TRUE,
  display.mode = "normal",
  test.mode = FALSE
)
}
\arguments{
\item{config_file}{TODO}

\item{config_dir}{TODO}

\item{host}{TODO}

\item{launch.browser}{TODO}

\item{port}{TODO}

\item{quiet}{TODO}

\item{display.mode}{TODO}

\item{test.mode}{TODO}
}
\description{
\code{web_app} returns an object of class \code{shinyApp} that can be passed to \code{shiny::runApp} to serve the IMLUI Web App. Most of the time you should not use \code{web_app} directly but instead use \code{\link[=serve_web_app]{serve_web_app()}}.
}
\seealso{
\code{\link[=serve_web_app]{serve_web_app()}}
}
