% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{wolframs_mouse_data}
\alias{wolframs_mouse_data}
\title{Wolframs Mouse Data}
\format{
A data.frame with 28 obs. of 1962 variables. Column 1 (drug)
describes the mouse's treatment (EMPA or H20) and column 2 (livers)
the mouses amount of livers (probably Sh means 1 and UNx means 2,
but not sure about that... TODO). Columns 3-1962 describe protein
levels:\preformatted{1. drug [Levels: EMPA, H2O]
2. livers [Levels: Sh, UNx]
3. sp|A2AJL3|FGGY_MOUSE [Range: 121816-332546]
4. sp|A2AKK5|ACNT1_MOUSE [Range: 156999-386092]
5. sp|A2ALS5|RPGP1_MOUSE [Range: 7587-96635]
6. sp|A2APT9|KLD7A_MOUSE [Range: 4224-41638]
7. sp|A2ARV4|LRP2_MOUSE [Range: 3855440-9426737]
8. sp|A2ASQ1|AGRIN_MOUSE [Range: 11294-109993]
9. sp|A2ATU0|DHTK1_MOUSE [Range: 42728-172820]
10. sp|A3KMP2|TTC38_MOUSE [Range: 195289-460792]
11. sp|B1AVD1|XPP2_MOUSE [Range: 39261-217552]
12. sp|B2RSH2|GNAI1_MOUSE [Range: 14384-219362]
13. sp|B9EJ86|OSBL8_MOUSE [Range: 93481-278556]
14. sp|D3Z7P3|GLSK_MOUSE [Range: 148542-325257]
15. sp|E9PVA8|GCN1_MOUSE [Range: 4658-218314]
16. sp|E9Q3L2|PI4KA_MOUSE [Range: 11597-160966]
17. sp|E9Q557|DESP_MOUSE [Range: 94059-1126995]
18. sp|E9Q612|PTPRO_MOUSE [Range: 22785-138597]
19. sp|E9Q634|MYO1E_MOUSE [Range: 21909-144416]
20. sp|G3X9C2|FBX50_MOUSE [Range: 56436-272938]
21. sp|G5E829|AT2B1_MOUSE [Range: 26444-81168]
22. sp|G5E8K5|ANK3_MOUSE [Range: 187914-613215]
23. sp|O08529|CAN2_MOUSE [Range: 56052-221873]
24. sp|O08547|SC22B_MOUSE [Range: 51032-189268]
25. sp|O08553|DPYL2_MOUSE [Range: 63367-902331]
}
}
\source{
\url{https://www.spang-lab.de/people/lbuck}
}
\usage{
wolframs_mouse_data
}
\description{
Protein profiles of mouses with different amounts of livers treated
with different drugs (H20 or EMPA).
}
\examples{
r$> wolframs_mouse_data[,1:5]
       drug livers sp|A2AJL3|FGGY_MOUSE sp|A2AKK5|ACNT1_MOUSE sp|A2ALS5|RPGP1_MOUSE
M29651  H2O     Sh               311029                225797                 22829
M29613  H2O     Sh               196401                296919                 25143
M29611  H2O     Sh               163090                312092                 11310
M29691  H2O     Sh               179499                386092                 11655
M29631  H2O     Sh               168107                275513                  7587
M29693  H2O    UNx               214390                229509                 19347
M29711  H2O    UNx               176614                321788                 30677
M29632  H2O    UNx               156334                232193                 16637
M29712  H2O    UNx               332546                220514                 48960
M29641 EMPA     Sh               180234                156999                 29859
M29662 EMPA     Sh               197679                319568                 18867
M29721 EMPA     Sh               235326                245869                 37981
M29621 EMPA     Sh               163605                269223                 31367
M29701 EMPA     Sh               181955                310191                 35756
M29622 EMPA    UNx               227559                306449                 17950
M29722 EMPA    UNx               275727                306962                 20389
M29623 EMPA    UNx               128854                380689                  8912
M29642 EMPA    UNx               239226                307414                 24241
M29672  H2O     Sh               219573                202160                 39417
M29671  H2O     Sh               121816                214691                 24686
M29652  H2O    UNx               291227                361385                 51186
M29731  H2O    UNx               252188                312156                 96635
M29692  H2O    UNx               205545                338143                 84372
M29681 EMPA     Sh               169352                246204                 17015
M29653 EMPA     Sh               180709                252661                 81127
M29682 EMPA     Sh               213966                284455                 24757
M29702 EMPA    UNx               171259                310694                 29816
M29661 EMPA    UNx               306009                315091                 45640
}
\keyword{datasets}
