% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/900_mocks.R
\name{mock_input_obj}
\alias{mock_input_obj}
\title{Return a shiny server input object (for Testing Purposes)}
\usage{
mock_input_obj(...)
}
\arguments{
\item{...}{Properties of returned input object}
}
\description{
Return a shiny server input object (for Testing Purposes)
}
\examples{
input <- mock_input_obj(x = 1, y = 2)
input$x == 1 && input$y == 2
}
