% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/php.R
\name{plot_predictions_histogram}
\alias{plot_predictions_histogram}
\title{Predictions Histogram Plot}
\usage{
plot_predictions_histogram(
  predictions,
  density_lines = TRUE,
  rug = TRUE,
  binwidth = NULL
)
}
\arguments{
\item{predictions}{as the name says (named list of named numeric vectors)}

\item{density_lines}{add density lines to the plot (logical)? (logical)}

\item{rug}{add rugs to the plot? (logical)}

\item{histogram}{add a histogram to the plot? (logical)}
}
\value{

}
\description{
Predictions Histogram Plot
}
\examples{
TODO
}
