% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{predict.numeric}
\alias{predict.numeric}
\title{Predict method for linear models provided as plain named vectors}
\usage{
\method{predict}{numeric}(b, X, fm = NULL)
}
\arguments{
\item{b}{Named numeric vector}

\item{X}{Numeric matrix with \code{names(b) \%in\% c("Intercept", colnames(X))}}
}
\value{

}
\description{
Predict method for linear models provided as plain named vectors
}
\examples{
TODO: remove this method and make `lamis_signature` an object with a own predict method
}
